chore(tests) move unit tests out of lib
